% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/architect.R
\docType{class}
\name{blueprint}
\alias{blueprint}
\title{Definition Blueprint (EXPERIMENTAL)}
\description{
Class \code{blueprint} is set as an attribute of the output from each successfully completed invocation of \code{\link{define}}().  Blueprints execute before the \code{...} arguments and are appended with each call.  The idea is to be able to recreate a series of operations by using the \code{blueprint} as a starting point with a compatible dataset (i.e., contains enough features to satisfy the blueprint).
}
\section{Slots}{

\describe{
\item{\code{schema}}{A list of \code{\link{define}} operation expressions}
}}

