% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchspace.R
\name{cleanUpX}
\alias{cleanUpX}
\title{Clean up Sampled Values for [`create_model_genomenet()`]}
\usage{
cleanUpX(x, num_targets)
}
\arguments{
\item{x}{(`list`)\cr
List of numeric hyperparameter values for [`create_model_genomenet()`].}

\item{num_targets}{(integer `numeric(1)`)\cr
Value to set the `num_targets` entry in the resulting `list` to.}
}
\value{
`list`: the converted hyperparameter values.
}
\description{
Even though some of the hyperparameters of the [`create_model_genomenet()`]
network are categorical, it is easier to use a urely numeric search space,
e.g. the one created by [`get_searchspace()`]. `cleanUpX()` converts
numeric values to the right type for [`create_model_genomenet()`].
}
