% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{define}
\alias{define}
\title{Define a Data Transformation Schematic}
\usage{
define(
  data = NULL,
  ...,
  keep.rownames = TRUE,
  blueprint = NULL,
  progress = FALSE
)
}
\arguments{
\item{data}{The input data (see 'Details')}

\item{...}{(\code{\link[rlang]{dots_list}}): Formulas for which the left-hand side (LHS) is an expression containing the operation, and the right-hand side (RHS) contains column names that form a grouping set for the operation (i.e., \code{<expression> ~ col_1 + col_2 + ...}):
\itemize{
\item{If the form \code{<LHS>~ .} is given, the LHS executes using all columns as the grouping set}
\item{If the form \code{<LHS>~ 1} is given, the LHS executes without grouping}
\item{If no LHS is given, the operation defaults to selection based on the RHS}
}}

\item{keep.rownames}{See \code{\link[data.table]{data.table}}}

\item{blueprint}{(EXPERIMENTAL) See \code{\link{blueprint}}. Blueprints are always evaluated first when provided: this makes it easy to "branch" off of a base dataset.}

\item{progress}{(logical | FALSE) Should a progress bar be shown during execution (passed to \code{\link[purrr]{map}})?}
}
\value{
The data modified
}
\description{
\code{define()} allows one to operate on data using one or more formula-based definitions. Data transformation and selection can be achieved with formulas or using standard \code{\link[data.table]{data.table}} syntax in a procedural manner with a single function call. Each operation operates on the previous result making \code{define()} a pipeline of operations in a compact framework.
}
\details{
\itemize{
\item{If \code{data} is a \code{\link[smart.data]{smart.data}} object, the taxonomical references to field names can be accessed by using the \code{use()} syntax in the right-hand side of the formula (e.g., \code{~use(term1, term2) + otherTerm1 + ...}).}
\item{When using formulas for \code{...} on an empty \code{\link[data.table]{data.table}} object, set the argument \code{data} to \code{NULL}; otherwise, you will get an error.}
}
}
