% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuhelpers.R
\name{estimateBatchSize}
\alias{estimateBatchSize}
\title{Estimate Batch Size to be used for GenomeNet Model}
\usage{
estimateBatchSize(
  x,
  defaultargs,
  tolerance = 0.05,
  downgrade = 0.9,
  upper.limit = NULL
)
}
\arguments{
\item{x}{(`list`)\cr
Arguments given to [`create_model_genomenet()`] to create the model to be tried.}

\item{defaultargs}{(`list`)\cr
Arguments given to [`deepG::trainNetwork`] to train the model. The following arguments
should not be given: `model` (as this will be the result of `create_model_genomenet(x)`),
`batch.size` (as this is determined by search), `epochs` (set to 2), `steps.per.epoch` (set to 2),
`output` (set to `none`), `validation.split` (set to 0.5), `max_samples` (set to `NULL`).}

\item{tolerance}{(`numeric(1)`)\cr
Relative tolerance.}

\item{downgrade}{(`numeric(1)`)\cr
Value to multiply the value of the largest known good model with before returning. This should
be < 1 and is recommended since training for many epochs sometimes uses more memory and saturates
the GPU later.}

\item{upper.limit}{(`numeric(1)`)\cr
Largest model to try. If not given, it is set to 400 if `x$maxlen` is > 1000, and 2048 otherwise.}
}
\value{
`numeric(1)`: Recommended, reasonably safe, batch size.
}
\description{
Try out batchsizes for models that evaluate a single batch.
If the model is too large (i.e. an error is generated), a larger batchsize is tried.
If the model fits in the GPU, a larger batchsize is tried.
If the relative difference between the largest known good model size, and the smallest
known bad model size, is less than `tolerance` (or if the absolute difference is 1),
the search stops.
The resultin proposed model size is then multiplied by `downgrade` as a safety measure,
since models too close to the memory limit may sometimes fail later.
The minimum value returned is 1.

Search happens in two stages: At first, models being tried are grown exponentially
between 2 and `upper.limit`, followed by binary search.

Binary search experiments are run on an external `callr`-session, to avoid crashes on the
GPU bringing down the master R session.
}
