% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuhelpers.R
\name{estimateStepsPerEpoch}
\alias{estimateStepsPerEpoch}
\title{Estimate Steps per Epoch to do in a Given Time}
\usage{
estimateStepsPerEpoch(time.per.epoch, defaultargs, gen.val, validationsteps)
}
\arguments{
\item{time.per.epoch}{(`numeric(1)`)\cr
Time in seconds that a single epoch should take.}

\item{defaultargs}{(`list`)\cr
Arguments of `trainNetwork`. Values that should not be present
are `epochs` (set to 1), `steps.per.epoch` (set to various values for
careful estimation), `output` (set to `none`), `validation.split`
(set directly), and `gen.val` (given to `estimateStepsPerEpoch()` directly).}

\item{gen.val}{(`function`)\cr
Validation generator. Since validation is sometimes switched off in the experiments
conducted, this should be given separately.}

\item{validationsteps}{(integer `numeric(1)`)\cr
Number of validation steps that are being performed. The time that a validation step
takes is estimated separately and taken into account when determining te number of steps
to make.}
}
\value{
`numeric(1)`: The recommended number of steps per epoch.
}
\description{
Estimates the number of steps that can be done approximately in a given time.
}
