% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuhelpers.R
\name{evaluatePerf}
\alias{evaluatePerf}
\title{Evaluate performance of a model fitting run}
\usage{
evaluatePerf(x, epochs_desired)
}
\arguments{
\item{x}{(`numeric`)\cr
Vector of validation values. Larger values should be better.}

\item{epochs_desired}{(`numeric(1)`)\cr
Target number of epochs used by the evaluation.
See also `getEpochsDesired` parameter of [`makeGenomeNetObjective()`].}
}
\value{
`numeric(1)` the performance value.
}
\description{
Performance is evaluated as the maximum of all validation values reached.
However, if very many epochs were evaluated, this has an optimistic bias
since validation performance also has a random component.
Performance for runs that have more validation values than 50, and more than
40% of `epochs_desired`, is therefore calculated as the second-highest
value of the last `0.2 * epochs_desired` validation values encountered.
}
