% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchspace.R
\name{get_searchspace}
\alias{get_searchspace}
\title{Get Search Space for Genome Sequence Models.}
\usage{
get_searchspace(maxlen, type)
}
\arguments{
\item{maxlen}{(integer `numeric(1)`)\cr
Input sequence length. This value determines the upper bounds of some
of the search intervals.}

\item{type}{(`character(1)`)\cr
One of `"gap"` or `"recurrent`. Used to include some of the model-type
specific search space components.}
}
\value{
a [`ParamSet`].
}
\description{
Create a search space over which the model created by
[`create_model_genomenet()`] can be tuned.

Even though some of the hyperparameters of the network are categorical,
this creates a purely numeric search space, which can be optimized over
more easily. It is therefore necessary to convert sampled values
using [`cleanUpX()`] before calling [`create_model_genomenet()`].
}
