% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuhelpers.R
\name{getGpuMemFree}
\alias{getGpuMemFree}
\title{Get free memory of GPU in use}
\usage{
getGpuMemFree()
}
\value{
`numeric(1)`.
}
\description{
Uses `nvidia-smi` and parses the result.
Only works when a single `GPU` is present or assigned via
`CUDA_VISIBLE_DEVICES`.
}
