% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_mapper.R
\docType{class}
\name{jmap-class}
\alias{jmap-class}
\alias{jmap}
\title{A Join Map}
\description{
Class \code{jmap} contains the results of a call to \code{\link{join_mapper}}
}
\section{Slots}{

\describe{
\item{\code{map}}{The map produced as a result of calling \code{\link{join_mapper}}}

\item{\code{env}}{The environment of assignment}
}}

