% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.reduce.R
\name{join.reduce}
\alias{join.reduce}
\title{Join-Reduce Multiple Datasets}
\usage{
join.reduce(
  join_map,
  out_name,
  x_names,
  i_names,
  filters = TRUE,
  dt_key,
  source_env = attr(join_map, "env"),
  assign_env = attr(join_map, "env"),
  clean = FALSE
)
}
\arguments{
\item{join_map}{The input join-map (see \code{\link{join.mapper}})}

\item{out_name}{The name of the output object}

\item{x_names}{(string[]) Names or REGEX patterns indicating the outer table: multiple values will be concatenated into a delimited string}

\item{i_names}{(string[]) Names or REGEX patterns indicating the inner table to be joined: multiple values will be concatenated into a delimited string}

\item{filters}{(expression[[]]) \code{data.table}-friendly expression to limit the rows of the output. A length-2 list indicates \emph{row} and \emph{column} expressions respectively.}

\item{dt_key}{(string[], symbol[]) Names that participate in creating the \code{\link[data.table]{key}} for the output}

\item{source_env}{The environment from which the input should be sourced.  If empty, it defaults to \code{attr(jmap, "env")}}

\item{assign_env}{The environment in which the output should be assigned.  If empty, it defaults to \code{attr(jmap, "env")}}

\item{clean}{(logical) \code{TRUE} indicates that the object should be removed from \code{env} before creating the output}
}
\value{
Combined datasets assigned to the designated environment
}
\description{
\code{join.reduce} leverages \code{\link[data.table]{data.table}} fast joins and \code{\link[purrr]{reduce}} from package \code{purr}
}
