% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitnessfun.R
\name{makeGenomeNetObjective}
\alias{makeGenomeNetObjective}
\title{Create an Objective Function to be Optimized with MBO}
\usage{
makeGenomeNetObjective(
  maxlen,
  type,
  getEpochsDesired,
  path,
  path.val,
  labels,
  plgpath,
  outdir
)
}
\arguments{
\item{maxlen}{(integer `numeric(1)`)\cr
Input sequence length of models to evaluate.}

\item{type}{(`character(1)`)\cr
One of `"gap"` or `"recurrent`.}

\item{getEpochsDesired}{(`function`)\cr
A `function` that maps a `numeric(1)`, indicating walltime in number of seconds, to an integer
valued `numeric(1)`, indicating the number of epochs that should approximately be evaluated.}

\item{path}{(`character(1)`)\cr
Training data path.}

\item{path.val}{(`character(1)`)\cr
Validation data path.}

\item{labels}{(`character`)\cr
Class labels.}

\item{plgpath}{(`character(1)`)\cr
Path of a preloaded generator that is read with [`readPLG()`].}

\item{outdir}{(`character(1)`)\cr
Output directory. The tensorboard output is put in the `tensorboard_opt/` subdirectory of this.}
}
\value{
`function`.
}
\description{
Creates the objective function that should be optimized.

The objective function automatically adjusts batch size to fit in a given GPU,
and adjusts epoch size so that an approximate target number of desired epochs can be
evaluated in the requested `walltime`.

The function should be evaluated inside its own R-session.
}
