% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multicoreGenerator.R
\name{multicoreGenerator}
\alias{multicoreGenerator}
\title{Create a multicore Generator}
\usage{
multicoreGenerator(
  gen.constructor,
  seed,
  n.workers = NULL,
  queue.depth = NULL,
  cache.size.mb = NULL,
  cluster.constructor = NULL,
  packages = "deepG"
)
}
\arguments{
\item{gen.constructor}{(`function`)\cr
A function that returns a generator. This function is called in each
process. It should take a single argument `seed` and return a generator.
This generator is itself a function with no arguments that returns a
new batch of data each time it is called.}

\item{seed}{(integer `numeric(1)`)\cr
The seed to use. Individual workers are seeded using `seed + 0:(n.workers - 1)`.}

\item{n.workers}{(integer `numeric(1)`)\cr
The number of workers to use. If `NULL`, the number of workers is
determined using `parallelly::availableCores()`.}

\item{queue.depth}{(integer `numeric(1)`)\cr
The number of batches to load in advance. Making this larger prevents
stalls due to variations in individual `gen()`-calls' runtimes (which
can happen due to variations in FASTA file sizes), but also increases
memory usage.
Defaults to `NULL`: Use `cache.size.mb` instead. Note that exactly one of
`queue.depth` and `cache.size.mb` must be given.}

\item{cluster.constructor}{(`cluster`)\cr
Function that creates the cluster to use. It must be a `parallel` or `parallelly` PSOCK cluster.
By default, a cluster is created using `parallelly::makeClusterPSOCK()`.}

\item{`cache.size.mb`}{(`numeric(1)`)\cr
Approximate memory usage of the queue datastructure on the master process in megabytes. This
only counts the content of batches, not the overhead of the queue, so actual
memory usage is slightly higher. This is independent of individual worker
process memory usage, which is determined by the generators used and will include
at least one batch of data.
Defaults to `NULL`: Use `queue.depth` instead.
Note that exactly one of `queue.depth` and `cache.size.mb` must be given.}
}
\description{
Launches a multicore generator that loads data from disk in parallel.
}
