% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbohelpers.R
\name{resetOptStateItersTerminator}
\alias{resetOptStateItersTerminator}
\title{Set MBO Control Object to Terminate after a Certain Number of Iterations}
\usage{
resetOptStateItersTerminator(opt.state, iters, save.on.disk.at)
}
\arguments{
\item{opt.state}{(`OptState`)\cr
The `OptState` to reset.}

\item{iters}{(integer `numeric(1)`)\cr
The (total) number of iterations to run. Note that this is not the number of iterations to run after the reset, but
the total number of iterations after which to terminate.}

\item{save.on.disk.at}{(integer `numeric`)\cr
The number of iterations after which to save the `OptState` to disk. Can be a vector of integers, in which case the
`OptState` is saved to disk after each of the given number of iterations.}
}
\value{
the modified `opt.state`.
}
\description{
Resets the termination state of a given `opt.state` to terminate after a certain number of iterations.
}
