% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbohelpers.R
\name{resetOptStateObjective}
\alias{resetOptStateObjective}
\title{Changes the Objective of an Optimization State}
\usage{
resetOptStateObjective(
  opt.state,
  fun,
  defaults = getDefaults(getParamSet(fun))
)
}
\arguments{
\item{opt.state}{(`OptState`)\cr
The `OptState` to modify.}

\item{fun}{(`smoof_function`)\cr
The new objective function.}

\item{defaults}{(list)\cr
A list of default values to use for the new parameters. If not given, each parameter's default (or `NA`, if not
present) is used.}
}
\value{
the modified `opt.state`.
}
\description{
Modifies the objective of a given `OptState`, taking care to update the `OptState`'s `ParamSet` using
[resetOptStateParamSet()].
}
