% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbohelpers.R
\name{resetOptStateParamSet}
\alias{resetOptStateParamSet}
\title{Enlarges the Search Space of an Optimization State}
\usage{
resetOptStateParamSet(opt.state, par.set, defaults = getDefaults(par.set))
}
\arguments{
\item{opt.state}{(`OptState`)\cr
The `OptState` to modify.}

\item{par.set}{(`ParamSet`)\cr
The new `ParamSet` to use.}

\item{defaults}{(list)\cr
A list of default values to use for the new parameters. If not given, each parameter's default (or `NA`, if not
present) is used.}
}
\value{
the modified `opt.state`.
}
\description{
Changes the search space to the given `ParamSet` and adds the required columns to the evaluation `OptPath`.
To fill in values, the 'default' value, or `NA`, is used if the `defaults` argument is not given.

Removing search space components is not supported.
}
