% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{to_rhs}
\alias{to_rhs}
\title{Convert to RHS Formula Terms}
\usage{
to_rhs(...)
}
\arguments{
\item{...}{\code{\link{rlang}{dots_list}} Strings or symbols to use}
}
\value{
A partial right-hand-side of a formula
}
\description{
\code{to_rhs} takes its input and creates an language object that can be included in the right-hand side of a formula object.
}
\examples{
res <- define(
	NULL
	, x = 1:10
	, y = x * 3
	, z = sum(y) ~ x
	, omega := 100
	)
print(res)
print(res <- define(res, ~!!to_rhs(z, omega)))
attributes(res)$blueprint
}
